Rename com.liferay.portal.apio to portal.headless to avoid split package with the older headless-apio module
