--- modules/apps/portal-apio/portal-apio-impl/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Portal Apio Implementation
-Bundle-SymbolicName: com.liferay.portal.apio.impl
-Bundle-Version: 1.0.0
