+++ modules/apps/portal-headless/portal-headless-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Headless API
+Bundle-SymbolicName: com.liferay.portal.headless.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.headless.context
