+++ modules/apps/portal-headless/portal-headless-api/src/main/resources/com/liferay/portal/headless/context/packageinfo
