+++ modules/apps/portal-headless/portal-headless-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Portal Headless Implementation
+Bundle-SymbolicName: com.liferay.portal.headless.impl
+Bundle-Version: 1.0.0
