+++ modules/apps/portal-headless/portal-headless-impl/src/main/java/com/liferay/portal/headless/internal/jaxrs/json/JSONMessageBodyWriter.java
@@ -1,0 +1,85 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.headless.internal.jaxrs.json;
+
+import com.fasterxml.jackson.databind.MapperFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+
+import java.io.IOException;
+import java.io.OutputStream;
+
+import java.lang.annotation.Annotation;
+import java.lang.reflect.Type;
+
+import javax.ws.rs.Produces;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.MultivaluedMap;
+import javax.ws.rs.ext.MessageBodyWriter;
+import javax.ws.rs.ext.Provider;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.jaxrs.whiteboard.JaxrsWhiteboardConstants;
+
+/**
+ * @author Ivica Cardic
+ */
+@Component(
+	property = JaxrsWhiteboardConstants.JAX_RS_EXTENSION + "=true",
+	service = MessageBodyWriter.class
+)
+@Produces(MediaType.APPLICATION_JSON)
+@Provider
+public class JSONMessageBodyWriter implements MessageBodyWriter<Object> {
+
+	public long getSize(
+		Object object, Class<?> clazz, Type genericType,
+		Annotation[] annotations, MediaType mediaType) {
+
+		return -1;
+	}
+
+	public boolean isWriteable(
+		Class<?> clazz, Type genericType, Annotation[] annotations,
+		MediaType mediaType) {
+
+		if (clazz == String.class) {
+			return false;
+		}
+
+		return _objectMapper.canSerialize(clazz);
+	}
+
+	public void writeTo(
+			Object object, Class<?> type, Type genericType,
+			Annotation[] annotations, MediaType mediaType,
+			MultivaluedMap<String, Object> multivaluedMap,
+			OutputStream outputStream)
+		throws IOException, WebApplicationException {
+
+		_objectMapper.writeValue(outputStream, object);
+
+		outputStream.flush();
+	}
+
+	private static final ObjectMapper _objectMapper = new ObjectMapper() {
+		{
+			configure(MapperFeature.SORT_PROPERTIES_ALPHABETICALLY, true);
+			enable(SerializationFeature.INDENT_OUTPUT);
+		}
+	};
+
+}
